JIT: Fold more always failing casts
Folding `x isinst B` where `x` is statically of type `A` is currently
done via a call to `compareTypesForCast`. If it returns
`TypeCompareState::Must` then we know that `A` can be cast to `B`, which
also means any class derived from `A` can be cast to `B`, so we can
always fold that. However, if it returns `TypeCompareState::MustNot` we
only learn something if we know `x` is exactly `A`; otherwise, if `x` is
more derived than `A`, it could still potentially be cast to `B`.

In cases where the type hierarchies are distinct, however, we can still
know that some type checks will fail. To detect the case add a query in
the opposite direction; that is, if `B` cannot be cast to `A`, then the
type hierarchies are distinct, so in that case we can fold it,
regardless of having exact knowledge. This also requires some explicit
checks for interface types, but perhaps we can move those to the EE
side.

Fix #97000