[mono][interp] Fix some leaks during compilation (#97143)
* [mono][interp] Stop trying to inflate signature

The target method and, implicitly, its signature are already inflated. This was allocating a new signature every time the method was called which was leaked. On some of the bigger tests suites, that heavily use generics, this can reduce the mem usage in the order of GBs.

* [mono][interp] Free mheader in case of inline failure

The header local types are not used anywhere so we can just free it.