Fix mscordac exports
The mscordbi depends on the PAL in mscordac and so it requires that all
functions it uses from coreclr PAL are exported by the libmscordac.
This got broken recently by a change in a header consumed by the mscordbi
that pulled in usage of two functions for UTF16 to UTF8 conversion into
the mscordbi.

This issue prevents debuggers using libmscordbi to initiate debugging.

This change fixes it by exporting those two functions from libmscordac.