Improve performance of Order{By}{Descending}(...).First/Last
When there's no ThenBy, we can take a more optimized path that uses the TKey's comparer directly.  We already have a fast path for this case that converts the O(n log n) operation into O(n), but it employs a comparer that's much more complicated, and as that comparer is used in the inner loop, it makes a meaningful difference.