JIT: capture all write barrier helper addresses for SPMI (#97535)
When varying CSEs we can sometimes alter the write barrier that is
needed. In particular if we CSE a heap address computation we may
lose track of the fact that an indir is writing to the heap, and so
change which write barrier needs to be used.

See #97534.

Even if that's fixed it seems like we still might change our minds
for various reasons, so when running under SPMI, just collect all
the possible write barrier helper addresses.