Replumb TensorPrimitives.ConvertToHalf/Single to be on operator plan
ConvertToHalf and ConvertToSingle contain their own vectorized algorithms. Now that we have a framework for handling other conversions, we can move the core of their implementations into operators and then have them share the same logic as the rest of the conversions. Then ConvertToHalf/Single just delegate to ConvertTruncating and exist only as legacy.