Fix S.T.Json's ArrayPool threshold
A few code paths hardcoded "current" knowledge of the ArrayPool's upper bound. But that bound changed in .NET 6, so now STJ is allocating large byte arrays in places it could be using pooling.  This fixes the limit.

Also, most STJ code paths that use the ArrayPool try to stackalloc instead, but a few paths don't. This fixes them to do so as well where possible.