JIT: Remove fgReplaceSwitchJumpTarget; increase usage of fgReplaceJumpTarget (#97664)
In preparation for improving the proliferation of edge likelihoods (as part of #93020), we should use fgReplaceJumpTarget where possible, so that we modify predecessor edges in fewer places. I plan on updating fgReplaceJumpTarget to always copy the likelihood of the old edge to the new edge as a follow-up to this change.