Patch libunwind to hotfix validation pipe creation race
A change in libunwind 1.8.0
https://github.com/libunwind/libunwind/pull/536 has introduced a race in
one of its commits
(https://github.com/libunwind/libunwind/commit/ec03043244082b8f552881ba9fb790aa49c85468)
This race causes relatively rare unwind failures when multiple threads race
for an initialization of a pipe that is used to verify that a memory address is
accessible.

This change reverts the change in the affected file. Once a proper fix
is made in the libunwind, this patch would be reverted and the real fix
applied.

This issue is currently blocking source build and the new exception
handling enabling by default.