JIT: Ensure edge likelihoods 1/N (#97740)
Start enforcing that all FlowEdges have likelihoods set (ideally, to the right value).

Update the profile checker to allow checking just for likelihoods. Enable this for all the initial phases up through and including fgAddInternal.

Fix missing flow edge issues by generally having whatever bit of code calls `fgAddRef` also ensure likelihood is set. For many cases these are unconditional flows so the value is just 1.0. Sometimes we can pass the old edge in and have`fgAddRef` itself do the work. And in a few cases (switches) things are a bit more subtle.

The general plan is to leave things like this for now until we've fixed likelihoods everywhere and are checking all phases. Then we'll go back and try and refactor `fgAddRef` to do more of this automatically.

Contributes to #93020