Devirtualize more based on whole program analysis (#97812)
Whole program view lets us figure out whether a virtual method has been overriden by another method, or whether it can be considered `final`. Add a pass to collect this information and pass it to codegen,

It kicks in less often than I expected (for linear enough hierarchies or not-too-much-derived types we could already do this with guarded devirtualization or type sealing), but I can still a bit over 100 method bodies affected by this in the Stage1 app.