[mono] Add runtime support for Swift calling convention on x64 and arm64 (#94764)
This commit introduces basic support for the Swift calling convention in the mini and interpreter backends on x64 and arm64 architectures. When a C# function calls a Swift library, it is expected that the self context is stored in a dedicated register. Similarly, when a Swift function returns, the error is supposed to be placed in a specific register.

---------

Co-authored-by: Matous Kozak <matouskozak@seznam.cz>
Co-authored-by: Jan Kotas <jkotas@microsoft.com>
Co-authored-by: Aaron Robinson <arobins@microsoft.com>
Co-authored-by: Jeremy Koritzinsky <jkoritzinsky@gmail.com>
Co-authored-by: Aleksey Kliger (λgeek) <akliger@gmail.com>