JIT: stopping preference for ML CSE heuristic
Some initial attempts to improve the modelling of "stopping early" when
doing CSEs. This adds a simplistic register pressure estimate modelled
on the one we have now, where we find the weight of the Nth tracked local
and use that as a reference weight for deciding when a CSE might start
becoming costly.

Contributes to #92915.