New nuget: Microsoft.NET.Runtime.WorkloadTesting.Internal
Nuget to package the workload testing targets, and tasks for internal
use. This is a non-shipping nuget which can be used for by other dotnet
repos to test workloads.

`eng/testing/workloads-testing.targets` is extracted into a
`WorkloadTesting.Core.targets`, and the `dotnet/runtime` repo specific
bits are left in `eng/testing/workloads-testing.targets`.