[infra] Run Xunit runtime tests on Apple mobile platforms
This commit adds Xunit runtime tests for Apple mobile platforms.

There are two groups of runtime tests: Xunit and legacy. Legacy tests expose the main method and cannot be run with AppleTestRunner. The Xunit tests are executed with AppleTestRunner. In both cases, it is necessary to pass a list of assemblies to run.

Tests that are not part of any merged testing groups are grouped by subcategories and bundled together.