Disable linker warning LNK4104 when building native AOT libraries (#98084)
link.exe hardcodes a list of [magic symbol names](https://learn.microsoft.com/en-us/cpp/error-messages/tool-errors/linker-tools-warning-lnk4104?view=msvc-170) that it thinks should not be generated into import library and emits warnings if they are.

We don't have a mechanism to set visibility of `UnmanagedCallersOnly` methods and I don't think we'd want to make a public API for that. But this means that users get a warning that is not particularly actionable. We could either replicate this magic list of symbol names into compiler and generate these as `PRIVATE` (but then how do people opt out if they don't actually want this as `PRIVATE`?), or we can just suppress the warning.

This PR suppresses the warning.