This is an incomplete implementation of a new randomized allocation sampling feature. This commit is
only to gather feedback on the direction and is not intended to be checked in.

The is a resumption of some of the original work that occured in https://github.com/dotnet/runtime/pull/85750
although this is a different implementation approach. The overall goal is to do lightweight allocation
sampling that can produce unbiased approximations of what was actually allocated. Our current ETW AllocationTick
sampling is periodic but not randomized. Extrapolating from the AllocationTick samples can lead to unbounded
estimation errors in theory and also substantial estimation error observed in practice.

This commit is primarily to get feedback on adding a new adjustable limit pointer on
the coreclr EE side of allocation contexts. This allows breaking out of the fast path allocation helpers at an
arbitrarily selected sampling point that need not align with the standard GC allocation
quantum for SOH. With sampling off the fast_helper_alloc_limit would always be
identical to alloc_limit, but when sampling is on we would occasionally have them
diverge.

The key changes are:
gcheaputilities.h - defines a new ee_alloc_context which augments the existing gc_alloc_context with the extra fast helper limit field
gcheaputilities.h and threads.h - replace the global and per-thread storage of gc_alloc_context with the expanded ee_alloc_context
  to store the new field.
jithelpers.cpp, jitinterfacex86.cpp, and amd64 asm stuff - refactors fast path alloc helpers to use the new limit field instead
gchelpers.cpp - Updated Alloc() function recognizes when fast helper limit was exceeded to do some kind of sampling callback
  and update the fast helper limit when needed.

This commit doesn't contain any logic that actually turns the sampling on, determines the limit values needed for sampling,
or issues callbacks for sampled objects. That would come later if folks think this is a reasonable approach to intercept
allocations.