Don't allow creating 1-byte dead basic blocks
Fixes #97758.

There's likely no substitution benefit from that (can't do anything meaningful in 1 byte), but this was causing the above issue: we had a mismatch for some IL patterns where we could optimize uninstantiated method bodies, but couldn't optimize instantiated ones because they ran into the abort at the bottom of this diff.