Shrink eh_frame section in naot executables (#98355)
`-fno-exceptions` is not enough to stop generating `.eh_frames`. We need to also pass `-fno-asynchronous-unwind-tables`.

Saves 50 kB on an app using WKS gc. Possibly more on SRV GC since that has two copies of the GC.

Note that we still get unwinding information for the debugger because `-g` is going to force generation of `.debug_frame` instead. But `.debug_frame` goes to the symbols file.