JIT: Unify preheader/exit weight calculation logic
The preheader/exit weight calculation is essentially the same logic, so
unify to use the logic from the exit weight calculation. This also fixes
a bug in the preheader weight calculation introduced by my recent change
where we sometimes mistakenly used the weight of the newly inserted
preheader instead of the header block to compute a likelihood.