Fix stepping into assembly thunks (#98401)
Stepping into assembly thunks didn't work with the new object writer. This fixes it.

For following program, trying to step into `ToString` (in an unoptimized version of this code) was resulting into a step over.

The issue is that the `ToString` call in unoptimized code looks like this:

```
00007FF71B8A27EA  lea         rax,[__VirtualCall_Object__ToString (07FF71B591007h)]
00007FF71B8A27F1  mov         qword ptr [rbp-18h],rax
00007FF71B8A27F5  mov         rcx,qword ptr [p]
00007FF71B8A27F9  mov         rax,qword ptr [rbp-18h]
00007FF71B8A27FD  cmp         dword ptr [rcx],ecx
00007FF71B8A27FF  call        rax
```

I.e. we call a `__VirtualCall_Object__ToString` assembly helper that does the actual call. The helper looks like this:

```
00007FF71B591007  mov         rax,qword ptr [rcx]
00007FF71B59100A  jmp         qword ptr [rax+18h]
```

Without debug information, the debugger was simply throwing up hands and stepping over this.

We do however generate debug information using the magic F00F00 and FEEFEE line numbers. These line numbers instruct the debugger to ignore the helper and step into the thing the helper calls. The new object writer was dropping this debug info on the floor.