JIT: Discard simple dominated CSEs that will never be defs (#98417)
Be a bit smarter when indexing CSEs if we see the same value twice in the same
basic block and the second one has more exceptions than the first one. In that
case the first will never be a candidate def for the second, so it is always
better to just consider the second one to be the def.

This is particularly relevant with passthrough nodes like `CASTCLASS(x, y)`,
where `CASTCLASS` gets the same VN as its first operand, but with more
exceptions on it.  Previously we would consider x to be the def; this naturally
considers the full `CASTCLASS(x, y)` to be the def, which promises more
exceptions.

One thing to note is that `GT_COMMA` would benefit from this handling too, but
it is already special cased such that we create a separate candidate for the
`GT_COMMA` (that includes the exception set of op1) and its op2 (that doesn't).
That would also be a potential way to improve the situation here, but it
requires special casing all such cases.

The fix here is very simplistic and doesn't handle many cases. In particular we
could have seen x previously in another basic block, in which case we run into
the same issue again. I think switching CSE to be dominator tree based and
generating multiple candidates instead would be a better general solution;
hopefully we can work towards that. If not, this at least alleviates some of the
problem.