JIT: Refactor Compiler::optRedirectBlock; remove BasicBlock::CopyTarget (#98526)
Part of #93020. When cloning a set of basic blocks (such as during loop cloning, finally cloning, etc.), the successors of the copied blocks may be different from the successors of their original counterparts, due to the successors being copied as well. We currently handle this as so:

* Create the block copies without initializing their successors.
* Copy the target(s) of the original block over with BasicBlock::CopyTarget; don't create any edges yet, as the new block's target(s) may change.
* In Compiler::optRedirectBlock, determine if the new block's target needs to be redirected to the copy of the original target, using a mapping of original blocks to their copies. We may or may not also initialize/update the successor edges for the new block.

This PR simplifies this process in preparation for updating BasicBlock to track its successors via FlowEdge pointers, instead of BasicBlock pointers. This is the new process:

* Create the block copies without initializing their successors.
* In Compiler::optRedirectBlock, for each successor of the original block, determine if the new block should use the successor, or if the new block should be redirected to a copy of that successor, using the block map. Either way, all of the new block's successor edges must be initialized here.