Fix initial context for HW exceptions in the new EH (#98605)
I have made a last minute change before creating the PR for the final
fixes for the new EH that turned out to fire an assert in REGDISPLAY
initialization. I was originally zeroing the whole context, but then
figured out that just zeroing IP is sufficient to make the stack walker
start at an explicit frame. However, there is one assert that also
checks SP for validity (or 0) that started to fire with that change.

This PR fixes it.