Fix bug in arm64 LazyMachState unwindFix bug in arm64 LazyMachStateunwind
When we were enabling the mac M1 couple of years ago, a bug has creeped
in that causes a race in lazy machine state unwinding. The
LazyMachState::setLazyStateFromUnwind was incorrectly modifying the
captured state, so in case multiple threads were racing for the unwind,
a thread that came at a time when part of the captured state was
modified and before the `_isValid` was set would start unwinding from a
wrong state (the partial unwound one), effectively unwinding twice.

I've hit this in CI for enabling the new exception handling, as it
results in an assert when comparing the resulting address of the
floating point registers unwind with the mach state.